requirements.txt
asynch==0.2.3
build==1.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
ciso8601==2.3.1
clickhouse-connect==0.7.0
clickhouse-driver==0.2.6
clickhouse-sqlalchemy==0.3.0
greenlet==2.0.2
idna==3.6
leb128==1.0.5
lz4==4.3.3
numpy==1.24.4
packaging==23.1
pandas==2.0.3
psycopg2-binary==2.9.7
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools-scm==8.0.2
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.2.0
zstandard==0.22.0
zstd==1.5.5.1
